rust/host: the mint executor -- mint.rs, Rust mints its own eras
ADR-0030-in-progress step 4, the biggest piece: a faithful,
function-for-function port of lib/hecksagain/adapters/driven/
postgres_era/lineage/{provisioning,mint_transaction,head_compiler,
resumable_backfill}.rb's combined behavior. Every SQL string is the
SAME text those files issue, gathered verbatim from the real source,
not re-derived from a description -- provisioning (hecks_eras,
hecks_era_texts, hecks_approvals, the journal table, RLS, the six
hecks_tr_* transform functions), the mint transaction (advisory lock,
idempotency check, watermark capture, the hecks_eras row, grants, the
RLS fence flip), head compilation (matview + head-snapshot + head view
per aggregate, over the full non-layered ancestor chain), and the
resumable chunked backfill (SAVEPOINT-nested, progress-tracked).

Two deliberate, documented narrowings from Ruby, both because this
crate has no bluebook parser (rust/parser, never linked into rust/
host per ADR 0012): era 1 is named EAGERLY at hold time rather than
lazily (Ruby's ensure_named! re-parses held text later; this crate
already has its shape hash in hand when holding, so there's nothing
to defer), and an existing held era with a NULL hash (one Ruby minted
but never named) is out of scope to mint forward from -- refused by
name, not silently stalled. Also always uses the full chain build,
never head_compiler.rb's own layered-matview optimization (a
performance shortcut over an already-correct slower path, not a
distinct correctness-bearing code path) -- deferred, not skipped.

storage_shape.rs's FORM_VERSION was i64; hecks_eras.canon_form is
Postgres int4 -- fixed to i32, caught immediately by the first real
INSERT this module ever issued.

THE PROOF THIS WHOLE EFFORT EXISTS FOR: rust_mints_both_eras_itself_
writes_and_reads_back_the_translated_data -- Rust mints era 1
(hold_first), writes two real records through the already-proven
generic journal::append_lineage_mutation, mints era 2 itself via a
hand-built rename edge (mint_era), and reads the correctly-translated
data back through journal::read_lineage_head_all. Zero Ruby process
anywhere in this test -- the first time anything in this codebase has
minted an era without Ruby. Plus two fast pure unit tests for the
multi-hop chain-naming logic (names_by_era/edge_chain) a single-edge
test can't exercise. All 3 new tests green; full rust/host suite
otherwise unchanged (same 10 pre-existing, unrelated dist/banking.wasm
failures as before this work, 41 passing).

+++ rust/host/src/storage_shape.rs
@@ -41,7 +41,7 @@ pub const LABEL_LENGTH: usize = 6;
 
 // storage_shape.rb's own FORM_VERSION — bump this in the SAME change
 // that alters `project`/`canonical`'s output below, on BOTH sides.
-pub const FORM_VERSION: i64 = 1;
+pub const FORM_VERSION: i32 = 1;
 
 pub fn mint_hash(ir: &Value) -> String {
     let digest = Sha256::digest(canonical(ir).as_bytes());
